test(rxMisc): Remove yDiff functionality
This was only being used as a private function for one midway
test. Instead of publishing this hidden function for all consumers,
provide it as a one-off function just for the floating header test.
If you find yourself needing this for your application, consider
defining it yourself from the source in this change set.